API : move isotonic regression out of linear_model
